+++ docs/changes.rst
@@ -17,6 +17,7 @@ Unreleased.
  - Added :meth:`Image.get_image_distortion() <wand.image.BaseImage.get_image_distortion>` method.
  - Fixed `QuantumType` allocation for 32-bit architectures using HDRI. [:issue:`518`]
  - Fixed `MagickSizeType` allocation for :meth:`ResourceLimits.set_resource_limit() <wand.resource.ResourceLimits.set_resource_limit>` and segfault with ``armv7l`` architecture. [:issue:`520`]
+ - Fixed :class:`~wand.color.Color` deallocation error on 32-bit architectures.
  - [TEST] Deprecated PDF format from test assets.
  - [TEST] Deprecated :class:`~wand.drawing.Drawing` text ``fx_wand`` fixture to improve parallel CI testing.
  - [TEST] Marked all ImageMagick-7 features skipped when running test suite with ImageMagick-6. [:issue:`522`]